fix: retry if we can't get room parameters
If we can't get room parameters when we restart everything, try it
again. My suspicion (going off of no logs, because why would I
actually have logs) is that we get an intermittent failures when
reading room parameters leading to rooms to fall over one by one.
I'm deploying this to test in prod, we'll see what happens.